feat(update):update ingestion script with sumarization

+++ app/ingestion/ingestion_event.py
@@ -4,6 +4,11 @@ from langchain_community.document_loaders.pdf import PyPDFLoader
 from langchain_community.document_loaders import TextLoader
 from langchain.text_splitter import RecursiveCharacterTextSplitter
 from langchain_openai import OpenAIEmbeddings
+from langchain.chains.combine_documents.stuff import StuffDocumentsChain
+from langchain.chains.llm import LLMChain
+from langchain_core.prompts import PromptTemplate
+from langchain.chains.summarize import load_summarize_chain
+from langchain_openai import ChatOpenAI
 from langchain_community.vectorstores import Pinecone as PineconeLangChain
 from pinecone import Pinecone
 import os
@@ -14,8 +19,10 @@ os.environ["LANGCHAIN_ENDPOINT"] = "https://api.smith.langchain.com"
 os.environ["LANGCHAIN_API_KEY"] = "ls__41665b6c9eb44311950da14609312f3c"
 
 pc = Pinecone(api_key=os.environ.get("PINECONE_API_KEY"))
-INDEX_NAME = "adaletgpt-events-data"
+INDEX_NAME = "adaletgpt-summarized-event"
 DATASET = "../../dataset/"
+MIN_FILE_SIZE = 5 * 1024  # 5KB in bytes
+
 
 def get_subfolders(root_folder):
     for item in os.listdir(root_folder):
@@ -44,12 +51,36 @@ def ingest_docs():
         embedding_doc(file_path)
 
 def embedding_doc(file_path):
+    if os.path.getsize(file_path) > MIN_FILE_SIZE:
+        summarize_legal_case(file_path=file_path)
+        loader = TextLoader(file_path, encoding='utf-8')
+        raw_documents = loader.load()
+        embeddings = OpenAIEmbeddings(model="text-embedding-3-large")
+        PineconeLangChain.from_documents(raw_documents, embeddings, index_name=INDEX_NAME)
+        print("Inserting doc:", file_path)
+        os.remove(file_path)
+
+def summarize_legal_case(file_path):
+    prompt_template = """Write a concise summary of the following legal case in turkish:
+    Legal Case: {text}\n
+    CONCISE SUMMARY:"""
+    prompt = PromptTemplate.from_template(prompt_template)
+
+    # Define LLM chain
+    llm = ChatOpenAI(temperature=0, model_name="gpt-4-1106-preview")
+    llm_chain = LLMChain(llm=llm, prompt=prompt)
+    
+    # Define StuffDocumentsChain
+    stuff_chain = StuffDocumentsChain(llm_chain=llm_chain, document_variable_name="text")
     loader = TextLoader(file_path, encoding='utf-8')
-    raw_documents = loader.load()
-    embeddings = OpenAIEmbeddings(model="text-embedding-3-large")
-    PineconeLangChain.from_documents(raw_documents, embeddings, index_name=INDEX_NAME)
-    print("Inserting doc:", file_path)
-    os.remove(file_path)
+    docs = loader.load()
+    response = stuff_chain.invoke({"input_documents":docs})
+
+    with open(file_path, 'w', encoding='utf-8' ) as file:
+        file.write(response["output_text"])
+        print("Summarizing docs where is bigger than 5kbyte:", file_path)
+
 
 if __name__ == "__main__":
     ingest_docs()
+
